move to py_expr_functor
